# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "轻记 · 本地概念默写",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/exams": {
      "post": {
        "tags": [
          "考试"
        ],
        "summary": "Create Exam",
        "operationId": "create_exam_api_exams_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamStart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "考试"
        ],
        "summary": "List Exams",
        "operationId": "list_exams_api_exams_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/{exam_id}": {
      "get": {
        "tags": [
          "考试"
        ],
        "summary": "Get Exam",
        "operationId": "get_exam_api_exams__exam_id__get",
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/answers/{answer_id}/submit": {
      "post": {
        "tags": [
          "考试"
        ],
        "summary": "Submit Answer",
        "operationId": "submit_answer_api_answers__answer_id__submit_post",
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/{exam_id}/results": {
      "get": {
        "tags": [
          "考试"
        ],
        "summary": "Exam Results",
        "operationId": "exam_results_api_exams__exam_id__results_get",
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/{exam_id}/retry": {
      "post": {
        "tags": [
          "考试"
        ],
        "summary": "Retry Exam",
        "operationId": "retry_exam_api_exams__exam_id__retry_post",
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/answers/{answer_id}/retry-grade": {
      "post": {
        "tags": [
          "考试"
        ],
        "summary": "Retry Grade",
        "operationId": "retry_grade_api_answers__answer_id__retry_grade_post",
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/answers/{answer_id}/error-types": {
      "patch": {
        "tags": [
          "考试"
        ],
        "summary": "Update Labels",
        "operationId": "update_labels_api_answers__answer_id__error_types_patch",
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "const": true,
            "title": "Accepted",
            "default": true
          },
          "answer_id": {
            "type": "integer",
            "title": "Answer Id"
          }
        },
        "type": "object",
        "required": [
          "answer_id"
        ],
        "title": "AcceptedResponse"
      },
      "AnswerResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "time_spent_ms": {
            "type": "integer",
            "title": "Time Spent Ms"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "graded",
              "failed"
            ],
            "title": "Status"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "merged_feedback": {
            "additionalProperties": true,
            "type": "object",
            "title": "Merged Feedback"
          },
          "model_knowledge_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Knowledge Notes"
          },
          "judges": {
            "items": {
              "$ref": "#/components/schemas/JudgeResponse"
            },
            "type": "array",
            "title": "Judges"
          },
          "disagreement": {
            "type": "boolean",
            "title": "Disagreement"
          },
          "merge_fallback": {
            "type": "boolean",
            "title": "Merge Fallback"
          },
          "grading_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Error"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "error_types": {
            "items": {
              "$ref": "#/components/schemas/NamedType"
            },
            "type": "array",
            "title": "Error Types"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "position",
          "term",
          "user_answer",
          "time_spent_ms",
          "submitted_at",
          "status",
          "definition",
          "accuracy",
          "completeness",
          "final_score",
          "merged_feedback",
          "model_knowledge_notes",
          "judges",
          "disagreement",
          "merge_fallback",
          "grading_error",
          "prompt_version",
          "model_name",
          "created_at",
          "error_types"
        ],
        "title": "AnswerResponse"
      },
      "AnswerSubmit": {
        "properties": {
          "user_answer": {
            "type": "string",
            "maxLength": 200000,
            "title": "User Answer"
          },
          "time_spent_ms": {
            "type": "integer",
            "maximum": 31536000000.0,
            "minimum": 0.0,
            "title": "Time Spent Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_answer",
          "time_spent_ms"
        ],
        "title": "AnswerSubmit"
      },
...
      "ExamListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ExamListResponse"
      },
      "ExamResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "type": "string",
            "title": "Folder Name"
          },
          "parent_session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Session Id"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "submitted_count": {
            "type": "integer",
            "title": "Submitted Count"
          },
          "graded_count": {
            "type": "integer",
            "title": "Graded Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "pass_threshold": {
            "type": "number",
            "title": "Pass Threshold"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "folder_name",
          "parent_session_id",
          "started_at",
          "finished_at",
          "total",
          "submitted_count",
          "graded_count",
          "failed_count",
          "pass_threshold",
          "answers"
        ],
        "title": "ExamResponse"
      },
      "ExamStart": {
        "properties": {
          "folder_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Folder Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "folder_id"
        ],
        "title": "ExamStart"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JudgeResponse": {
        "properties": {
          "judge_index": {
            "type": "integer",
            "title": "Judge Index"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness"
          },
          "raw_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Json"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "judge_index",
          "success",
          "accuracy",
          "completeness",
          "raw_json",
          "raw_text",
          "error"
        ],
        "title": "JudgeResponse"
      },
      "LabelUpdate": {
        "properties": {
          "error_type_ids": {
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0.0
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Error Type Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error_type_ids"
        ],
        "title": "LabelUpdate"
      },
...
      "NamedType": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "NamedType"
      },
...
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "time_spent_ms": {
            "type": "integer",
            "title": "Time Spent Ms"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "graded",
              "failed"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "position",
          "term",
          "user_answer",
          "time_spent_ms",
          "submitted_at",
          "status"
        ],
        "title": "QuestionResponse"
      },
      "ResultsResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "type": "string",
            "title": "Folder Name"
          },
          "parent_session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Session Id"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "submitted_count": {
            "type": "integer",
            "title": "Submitted Count"
          },
          "graded_count": {
            "type": "integer",
            "title": "Graded Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "pass_threshold": {
            "type": "number",
            "title": "Pass Threshold"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "folder_name",
          "parent_session_id",
          "started_at",
          "finished_at",
          "total",
          "submitted_count",
          "graded_count",
          "failed_count",
          "pass_threshold",
          "answers"
        ],
        "title": "ResultsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}